+++ src/model/initialization.rs
@@ -1,4 +1,4 @@
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 
 use itertools::izip;
 use rand::prelude::ThreadRng;
@@ -7,12 +7,12 @@ use rand::Rng;
 use super::{Globals, Particle};
 
 pub fn generate_particle(rng: &mut ThreadRng, g: &Globals) -> Particle {
-    let x: Vec<f32> = izip!(&g.bounds_min, &g.bounds_max)
-        .map(|(min, max)| (rng.gen::<f32>() - 0.5) * (max - min))
+    let x: Vec<f64> = izip!(&g.bounds_min, &g.bounds_max)
+        .map(|(min, max)| (rng.gen::<f64>() - 0.5) * (max - min))
         .collect();
-    let v: Vec<f32> = (&x)
+    let v: Vec<f64> = (&x)
         .iter()
-        .map(|_| f32::sin(rng.gen::<f32>() * 2.0 * PI) * 10.0)
+        .map(|_| f64::sin(rng.gen::<f64>() * 2.0 * PI) * 10.0)
         .collect();
     Particle {
         position: x,
