+++ src/model/timestep.rs
@@ -2,7 +2,7 @@ use itertools::izip;
 
 use super::{Globals, Particle};
 
-fn step_1d(x: f32, v: f32, lower: f32, upper: f32, h: f32) -> (f32, f32) {
+fn step_1d(x: f64, v: f64, lower: f64, upper: f64, h: f64) -> (f64, f64) {
     let x = x + h * v;
     if x < lower {
         (lower, -v)
@@ -13,7 +13,7 @@ fn step_1d(x: f32, v: f32, lower: f32, upper: f32, h: f32) -> (f32, f32) {
     }
 }
 
-pub fn step(p: &Particle, h: f32, g: &Globals) -> Particle {
+pub fn step(p: &Particle, h: f64, g: &Globals) -> Particle {
     let (x, v) = izip!(&p.position, &p.velocity, &g.bounds_min, &g.bounds_max)
         .map(|(x, v, min, max)| step_1d(*x, *v, *min, *max, h))
         .unzip();
